# api/openapi.yaml
openapi: 3.1.0
info:
  title: dns-entree HTTP API
  version: 0.1.0-alpha
  summary: Stateless HTTP facade over the dns-entree library.
  description: |
    dns-entree is a provider-agnostic DNS automation library. This HTTP API
    is a thin, stateless wrapper that exposes the same operations as the
    `entree` CLI so callers in any language can use it via plain JSON.

    Deployment model: run behind a reverse proxy (Caddy, nginx, Traefik,
    Cloudflare Access, Tailscale, etc.) that handles TLS termination and
    authentication. The server itself is HTTP-only and unauthenticated.
    See docs/http-api.md for details.

    Credentials are passed per-request via X-Entree-* headers. The server
    holds nothing between requests. Two endpoints accept credentials in the
    JSON body instead of headers because they touch two providers at once
    (POST /v1/migrate, POST /v1/zone/import) or carry an RSA private key
    that must not appear in proxy access logs (POST /v1/dc/apply-url).

    Every successful response uses the envelope:
        { "ok": true, "schema_version": 1, "data": <payload> }
    Every error response uses:
        { "ok": false, "schema_version": 1, "error": { "code": "...", "message": "...", "details": {...} } }
servers:
  - url: http://localhost:8080
    description: Local server (default bind address)
...
components:
  securitySchemes:
    cloudflareToken:
      type: apiKey
      in: header
      name: X-Entree-Cloudflare-Token
      description: Cloudflare API token (zone-scoped). Used when X-Entree-Provider is `cloudflare`.
    awsAccessKeyId:
      type: apiKey
      in: header
      name: X-Entree-AWS-Access-Key-Id
      description: AWS access key id. Used with `route53` provider; pair with X-Entree-AWS-Secret-Access-Key.
    awsSecretAccessKey:
      type: apiKey
      in: header
      name: X-Entree-AWS-Secret-Access-Key
      description: AWS secret access key. Used with `route53` provider.
    godaddyKey:
      type: apiKey
      in: header
      name: X-Entree-GoDaddy-Key
      description: GoDaddy API key. Used with `godaddy` provider; pair with X-Entree-GoDaddy-Secret.
    godaddySecret:
      type: apiKey
      in: header
      name: X-Entree-GoDaddy-Secret
      description: GoDaddy API secret.
    gcdnsServiceAccount:
      type: apiKey
      in: header
      name: X-Entree-GCDNS-Service-Account-JSON
      description: Base64-encoded Google Cloud DNS service account JSON. Used with `google_cloud_dns` provider.
    providerSlug:
      type: apiKey
      in: header
      name: X-Entree-Provider
      description: Provider slug (`cloudflare`, `route53`, `godaddy`, `google_cloud_dns`). Required on endpoints that need a provider when more than one credential set is configured.
    bearerAccessToken:
      type: http
      scheme: bearer
      description: Per-migration capability token issued by POST /v1/migrate/preview. Required on all /v1/migrate/{id}/* routes.
...
  schemas:
    Envelope:
      type: object
      required: [ok, schema_version]
      properties:
        ok: { type: boolean }
        schema_version: { type: integer, enum: [1] }
        data:
          description: Payload (success only). Shape varies per endpoint.
        error:
          $ref: '#/components/schemas/ErrorBody'
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - MISSING_CREDENTIALS
            - BAD_REQUEST
            - PROVIDER_ERROR
            - VERIFY_MISMATCH
            - TEMPLATE_NOT_FOUND
            - RATE_LIMIT_EXCEEDED
            - INTERNAL
        message: { type: string }
        details:
          type: object
          additionalProperties: true
...
    Record:
      type: object
      required: [type, name, content]
      properties:
        type:    { type: string }
        name:    { type: string }
        content: { type: string }
        ttl:     { type: integer }
        prio:    { type: integer }
...
    Credentials:
      type: object
      description: Per-provider credential payload. Mutually exclusive sub-objects keyed by provider slug.
      additionalProperties: true
    MigrateRequest:
      type: object
      required: [domain, source_provider, target_provider]
      properties:
        domain:             { type: string }
        source_provider:    { type: string }
        target_provider:    { type: string }
        source_credentials: { $ref: '#/components/schemas/Credentials' }
        target_credentials: { $ref: '#/components/schemas/Credentials' }
        labels:             { type: array, items: { type: string } }
        dry_run:            { type: boolean }
    MigrationReport:
      type: object
      properties:
        domain:        { type: string }
        dry_run:       { type: boolean }
        records_total: { type: integer }
        records_ok:    { type: integer }
        records_fail:  { type: integer }
        errors:        { type: array, items: { type: string } }
    Zone:
      type: object
      properties:
        domain:  { type: string }
        records: { type: array, items: { $ref: '#/components/schemas/Record' } }
...
    StoredMigration:
      type: object
      description: Persisted migration row. `credential_blob` and `access_token` are always redacted on the wire.
      properties:
        id:                     { type: string }
        tenant_id:              { type: string }
        status:                 { type: string, enum: [preview, applying, awaiting_ns_change, verifying, complete, failed, expired] }
        domain:                 { type: string }
        source_slug:            { type: string }
        target_slug:            { type: string }
        preview:                { $ref: '#/components/schemas/Zone' }
        target_zone_id:         { type: string }
        target_nameservers:     { type: array, items: { type: string } }
        ns_change_instructions: { type: string }
        error_message:          { type: string }
        expires_at:             { type: string, format: date-time }
        version:                { type: integer }
        created_at:             { type: string, format: date-time }
        updated_at:             { type: string, format: date-time }
        credential_blob:        { type: string, description: "Always null on the wire (redacted)." }
        access_token:           { type: string, description: "Always empty on the wire (redacted)." }
...
  responses:
    Error400:
      description: Client error (bad request, missing credentials).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    Error404:
      description: Not found (template, domain).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
...
    Error500:
      description: Internal or provider failure.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
paths:
...
  /v1/migrate:
    post:
      tags: [migration]
      operationId: postMigrate
      deprecated: true
      summary: (Deprecated) Synchronous one-shot migrate
      description: |
        DEPRECATED. Use POST /v1/migrate/preview + POST /v1/migrate/{id}/apply
        + POST /v1/migrate/{id}/verify. Responses include `Deprecation: true`
        and `Sunset` headers. Will be removed after v0.2.0.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MigrateRequest' }
      responses:
        "200":
          description: migration report
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/MigrationReport' }
        "400": { $ref: '#/components/responses/Error400' }
        "500": { $ref: '#/components/responses/Error500' }
    get:
      tags: [migration]
      operationId: listMigrations
      summary: List stateful migrations
      parameters:
        - in: query
          name: tenant_id
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: list of migrations (credentials + access tokens redacted)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/StoredMigration' }
        "500": { $ref: '#/components/responses/Error500' }
...
  /v1/migrate/{id}:
    get:
      tags: [migration]
      operationId: getMigration
      summary: Fetch a stateful migration row (credentials + access token redacted)
      security:
        - bearerAccessToken: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: stored migration
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/StoredMigration' }
        "401": { $ref: '#/components/responses/Error400' }
        "404": { $ref: '#/components/responses/Error404' }
        "410": { $ref: '#/components/responses/Error400' }
    delete:
      tags: [migration]
      operationId: deleteMigration
      summary: Delete a stateful migration row
      security:
        - bearerAccessToken: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        "401": { $ref: '#/components/responses/Error400' }
        "404": { $ref: '#/components/responses/Error404' }
...